+++ src/apps/events/purpleserver/events/tests/__init__.py
@@ -1,0 +1,6 @@
+from purpleserver.manager.tests.test_addresses import *
+from purpleserver.manager.tests.test_parcels import *
+from purpleserver.manager.tests.test_shipments import *
+from purpleserver.manager.tests.test_trackers import *
+from purpleserver.manager.tests.test_custom_infos import *
+from purpleserver.manager.tests.test_pickups import *
