+++ src/apps/manager/purpleserver/manager/__init__.py
@@ -1,2 +1,1 @@
-
 default_app_config = 'purpleserver.manager.apps.ManagerConfig'
