+++ src/apps/manager/purpleserver/manager/apps.py
@@ -3,7 +3,3 @@ from django.apps import AppConfig
 
 class ManagerConfig(AppConfig):
     name = 'purpleserver.manager'
-
-    def ready(self):
-        from purpleserver.manager import jobs
-        jobs.start_schedulers()
